test: extract provider usage load coverage
